# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: C4 Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /api/admin
...
paths:
...
  /groups:
    post:
      operationId: PostGroups
      tags: [groups]
      summary: 创建分组（平台容量池；key 为独立表，由用户面 /user/keys 创建）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupCreate' }
      responses:
        '200':
          description: 创建后的分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetGroups
      tags: [groups]
      summary: 分组列表（分页/筛选/排序）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: name, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 分组列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/batch-delete:
    post:
      operationId: PostGroupsBatchDelete
      tags: [groups]
      summary: 批量删除分组（事务，全成或全败）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeleteBody' }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeleteResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/batch-update:
    post:
      operationId: PostGroupsBatchUpdate
      tags: [groups]
      summary: 批量更新分组（fields 为任意字段子集）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchUpdateGroupsBody' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchUpdateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetGroupsId
      tags: [groups]
      responses:
        '200':
          description: 分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutGroupsId
      tags: [groups]
      summary: 更新分组；携带 upstream_members 时与分组策略同事务替换
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupCreate' }
      responses:
        '200':
          description: 更新后分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeleteGroupsId
      tags: [groups]
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /groups/{id}/assignments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetGroupsIdAssignments
      tags: [groups]
      summary: 读取组的授予用户与专属倍率（platform_admin；与 PUT 对称，供前端预填充与安全全量写回）
      responses:
        '200':
          description: 该组当前的授予 user_id 列表与专属倍率
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupAssignmentsResponse' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutGroupsIdAssignments
      tags: [groups]
      summary: 设置组的授予用户（platform_admin；替换语义：未列出即撤销，空数组 = 清空）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupAssignmentsBody' }
      responses:
        '200':
          description: 已生效的授予 user_id 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupAssignmentsResponse' }
        default:
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean }
...
    GroupListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Group' }
    GroupVisibility:
      type: string
      enum: [public, private]
    GroupProtocolConvert:
      type: string
      enum: [auto, chat_to_resp, mess_to_resp, resp_to_mess, chat_to_mess]
      description: 分组级协议转换模式。auto = 自动协商（优先原生协议，再按固定优先级尝试转换）；其余值为手动方向。off 不在枚举内——不转换 = 空数组（见 protocol_convert 字段说明）
    GroupAssignmentsBody:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          items: { type: integer, format: int64 }
          description: 替换语义：完整授予列表（未列出即撤销；空数组 = 清空）
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
            minimum: 0
            maximum: 10
          description: 可选：user_id → 该用户在该组的专属价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算）。仅对 user_ids 中列出的用户生效；null = 清除为未设置（回退组倍率）；未列出的用户沿用当前值
    GroupAssignmentsResponse:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          items: { type: integer, format: int64 }
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
          description: 该组当前各用户的专属价格倍率（正常值；null/缺省 = 未设置 → 用组倍率）
...
    GroupCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        visibility: { $ref: '#/components/schemas/GroupVisibility' }
        routing_mode: { $ref: '#/components/schemas/GroupRoutingMode' }
        allowed_models:
          type: array
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 200 }
          description: 分组允许的模型白名单；空数组表示使用成员实际支持的模型交集
        upstream_members:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/GroupUpstreamInput' }
          description: 上游池成员。POST 在 routing_mode=upstreams 时与分组策略同事务创建；PUT 携带时与分组策略同事务替换，切回 accounts 时传空数组可清理旧成员
        price_multiplier: { type: number, format: double, minimum: 0, maximum: 10, nullable: true, description: 价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算——存储 15000 ↔ 显示 1.5）。缺省/null = 不设置（×1）；显式 0 = 免费组；PUT 显式写（含 0） }
        protocol_convert:
          type: array
          items: { $ref: '#/components/schemas/GroupProtocolConvert' }
          description: 协议转换模式。默认推荐 [auto]，优先原生协议并在缺少原生路由时自动选择转换；也可传单个或多个手动方向。空数组/缺省 = off = 不转换；auto 必须单独出现。PUT 语义：缺省（null/省略）= 保持原值；显式空数组 = 清空既有方向
...
    Group:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        Visibility: { $ref: '#/components/schemas/GroupVisibility' }
        RoutingMode: { $ref: '#/components/schemas/GroupRoutingMode' }
        PriceMultiplier: { type: number, format: double, description: 价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算——存储 15000 ↔ 显示 1.5） }
        AllowedModels:
          type: array
          items: { type: string }
        ProtocolConvert:
          type: array
          items: { $ref: '#/components/schemas/GroupProtocolConvert' }
          description: 协议转换模式（auto = 自动协商；空数组 = off = 不转换；手动方向按客户端格式命中）
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表/消费路径过滤已删；GET 单个可查已删项） }
    GroupRoutingMode:
      type: string
      enum: [accounts, upstreams]
      description: accounts 使用现有账号池；upstreams 使用本组上游成员
    GroupUpstreamInput:
      type: object
      required: [upstream_id]
      properties:
        upstream_id: { type: integer, format: int64, minimum: 1 }
        priority: { type: integer, minimum: 0, maximum: 100000, default: 0 }
        weight: { type: integer, minimum: 1, maximum: 10000, default: 100 }
        max_concurrency: { type: integer, minimum: 1, maximum: 100000, default: 8 }
        enabled: { type: boolean, default: true }
...
    BatchDeleteBody:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
...
    GroupPatch:
      type: object
      properties:
        name: { type: string }
        visibility: { $ref: '#/components/schemas/GroupVisibility' }
...
    BatchUpdateGroupsBody:
      type: object
      required: [ids, fields]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
        fields: { $ref: '#/components/schemas/GroupPatch' }
    BatchDeleteResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer }
    BatchUpdateResponse:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
...